Make the makefile totally explicit
I'm not sure there's a positive benefit to making the makefile this
explicit, certainly not for such a simple program, but at least this way
I can be sure about what steps are used to create each object file.

Also includes updates to the install command, and the addition of an
uninstall command.