Update unix_makefile for conciseness
This is not the maximum amount of brevity, and there's one glaring
inconvenience (the C file for an object file's recipe must be the first
argument) but I'm just exploring some options.

The general strategy is to (a) update VPATH to search for the files in
the right places, and (b) update the default build rules so that object
files get put in the right locations.